+++ src_java/mob/MobPoint.java
@@ -1,30 +1,33 @@
 package mob;
+
+
+
 import utils.Util;
 
 public class MobPoint {
 
     public final Mob mob;
-    public int hp;
-    public int maxHp;
-    public int dame;
+    public long hp;
+    public long maxHp;
+    public long dame;
 
     public MobPoint(Mob mob) {
         this.mob = mob;
     }
 
-    public int getHpFull() {
+    public long getHpFull() {
         return maxHp;
     }
 
-    public void setHpFull(int hp) {
+    public void setHpFull(long hp) {
         maxHp = hp;
     }
 
-    public int gethp() {
+    public long gethp() {
         return hp;
     }
 
-    public void sethp(int hp) {
+    public void sethp(long hp) {
         if (this.hp < 0) {
             this.hp = 0;
         } else {
@@ -32,8 +35,8 @@ public class MobPoint {
         }
     }
 
-    public int getDameAttack() {
-        return this.dame != 0 ? this.dame + Util.nextInt(-(this.dame / 100), (this.dame / 100))
+    public long getDameAttack() {
+        return this.dame != 0 ? this.dame + Util.nextLong(-(this.dame / 100), (this.dame / 100))
                 : this.getHpFull() * Util.nextInt(mob.pDame - 1, mob.pDame + 1) / 100
                 + Util.nextInt(-(mob.level * 10), mob.level * 10);
     }
